Move GetVersion to SDLEngine
