+++ src/sandbox/main.cpp
@@ -17,23 +17,10 @@ using window_ptr = std::unique_ptr<SDL_Window, deleter<SDL_Window, SDL_DestroyWi
 int main(int argc, char** argv) {
     SDL_Init(SDL_INIT_VIDEO);
 
-    window_ptr window{SDL_CreateWindow("Hello, World!", 800, 600, SDL_WINDOW_RESIZABLE)};
-    SDL_Surface* surface = SDL_GetWindowSurface(window.get());
+    int version = SDL_GetVersion();
 
-    SDL_FillSurfaceRect(surface, nullptr, SDL_MapSurfaceRGB(surface, 0xFF, 0xFF, 0xFF));
-
-    SDL_UpdateWindowSurface(window.get());
-
-    bool running = true;
-    SDL_Event e;
-
-    while (running) {
-        while (SDL_PollEvent(&e)) {
-            if (e.type == SDL_EventType::SDL_EVENT_QUIT) {
-                running = false;
-            }
-        }
-    }
+    std::cout << "SDL version: " << SDL_VERSIONNUM_MAJOR(version) << "." << SDL_VERSIONNUM_MINOR(version) << "." << SDL_VERSIONNUM_MICRO(version) << std::endl;
+    std::cout << "Hello, Aquila in C++!" << std::endl;
 
     SDL_QuitSubSystem(SDL_INIT_VIDEO);
     SDL_Quit();
